refactor: reorganize project structure for better maintainability
🏗️ Structure Changes:
- Created core/ directory for core systems
- Created scripts/ directory for auxiliary tools
- Kept main entry points in root

📁 Files Moved:
Core Systems → core/:
- logger.ts → core/logger.ts
- rlhf-system.ts → core/rlhf-system.ts

Auxiliary Scripts → scripts/:
- rlhf-autofix.ts → scripts/rlhf-autofix.ts
- rlhf-dashboard.ts → scripts/rlhf-dashboard.ts
- rollback-manager.ts → scripts/rollback-manager.ts

Entry Points (remain in root):
- execute-steps.ts (main executor)
- validate-template.ts (main validator)

🔧 Updates:
- Updated all imports to use new paths
- Updated package.json scripts with correct paths
- Added rlhf:dashboard and rollback scripts

✅ Benefits:
- Cleaner root directory with only entry points
- Logical organization: core/, scripts/, utils/
- Clear separation between main tools and auxiliaries
- Easier to navigate and understand
- No functionality lost, just better organized
